# docs/swagger.yaml
info:
  contact: {}
paths:
  /authenticate:
    post:
      consumes:
      - application/json
      description: It will authenticate an user from provided information
      operationId: authenticate-user
      parameters:
      - description: Username of the user
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: Password of the user
        in: body
        name: password
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The authentication is complete and access token provided in
            http only cookies with key 'auth-token'
          schema:
            type: boolean
        "210":
          description: Wrong authentication information provided
          schema:
            type: boolean
        "211":
          description: The siteGroup doesn't exists
          schema:
            type: boolean
      summary: Authenticate an user
  /authorize:
    get:
      consumes:
      - application/json
      description: It will authorize an user with jwt token from http only cookies
        with key name 'auth-token'
      operationId: authorize-user
      responses:
        "200":
          description: The user is authorized
          schema:
            type: boolean
        "210":
          description: Wrong jwt token
          schema:
            type: boolean
      summary: Authorize an user
  /register:
    post:
      consumes:
      - application/json
      description: It will register an user with provided information
      operationId: register-user
      parameters:
      - description: Username of the user
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: Password of the user
        in: body
        name: password
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The registration is completed
          schema:
            type: boolean
        "210":
          description: The user is already registered
          schema:
            type: boolean
        "211":
          description: The siteGroup doesn't exists
          schema:
            type: boolean
      summary: Register an user
swagger: "2.0"